Add files via upload
When i learned a bit of css, i made this sliding text display for my friend's birthday. Yes, i know i can make normal sliding text with css but this one has endless possibilities like each of the white box represents a pixel in the display and it can be set to any color, time frame and it is basically to explore more possibilities with css. I designed this display and all of the code is original. This single animation is a demonstration of how a couple of (331 to be accurate) animations can be used in sync to create a single beautiful animation. (By the way, that is his nickname lol)